refactor: get_address_from_message_and_signature function to drop repeated code
